pyproject.toml
[project]
name = "graphrag-chainlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.10.1",
    "chromadb>=1.5.5",
    "groq>=1.1.2",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.6",
    "neo4j>=6.1.0",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "tqdm>=4.67.3",
]
